refactor(habitat): normalize fix module model kinds
Collapse the fix module pattern/apply internals out of nested patterns/transactions buckets and into explicit model/dto, model/policy, and model/repositories kinds. DTO imports now come from the DTO barrel, behavior from policy, and worktree observation from repositories.

Validated with package typecheck, focused fix/pattern tests, service module shape, Grit pattern fixtures, Nx boundaries, boundary taxonomy, router import scan, and diff hygiene.
