+++ tools/habitat-harness/src/service/modules/fix/model/dto/index.ts
@@ -1,0 +1,4 @@
+export * from "./pattern-apply.schema.js";
+export * from "./pattern-management.schema.js";
+export * from "./shared.schema.js";
+export * from "./transaction-input.schema.js";
