+++ tools/habitat-harness/src/service/modules/fix/model/dto/pattern-management.schema.ts
