+++ tools/habitat-harness/src/service/modules/fix/model/policy/pattern-artifact-paths.policy.ts
