+++ tools/habitat-harness/src/service/modules/fix/model/policy/pattern-rule-reference.policy.ts
@@ -1,5 +1,5 @@
-import type { PatternRuleReferenceInput } from "./schema.js";
-import type { PatternRulePackReferenceInput } from "./validation.js";
+import type { PatternRuleReferenceInput } from "../dto/pattern-management.schema.js";
+import type { PatternRulePackReferenceInput } from "./pattern-validation.policy.js";
 
 export function patternRuleReferenceFromRule(
   rule: PatternRulePackReferenceInput
