--- tools/habitat-harness/src/service/modules/fix/model/policy/patterns/index.ts
@@ -1,90 +1,0 @@
-export {
-  applyAdmittedState,
-  diagnosticAdmittedState,
-} from "./admission.js";
-export {
-  activeApplyTransactionInputs,
-  applyTransactionInputsFromRuleFacts,
-  defaultApplyAdmissions,
-} from "./apply-admissions.js";
-export {
-  patternCandidateRoot,
-  patternManifestPath,
-  patternManifestRoot,
-  patternManifestSchemaVersion,
-} from "./paths.js";
-export { patternAdmissionRefusal } from "./refusal.js";
-export { patternRuleReferenceFromRule } from "./rule-reference.js";
-export {
-  type ApplyAdmission,
-  ApplyAdmissionSchema,
-  type ApplyDryRunCommand,
-  ApplyDryRunCommandSchema,
-  ApplyDryRunOutputModeSchema,
-  ApplyPatternPathSchema,
-  type ApplyTransactionInput,
-  ApplyTransactionInputSchema,
-  type CandidateHandoff,
-  CandidateHandoffSchema,
-  type CandidatePatternManifest,
-  CandidatePatternManifestSchema,
-  type DiagnosticAdmission,
-  DiagnosticAdmissionSchema,
-  type PatternAdmissionRefusal,
-  type PatternAdmissionRefusalReason,
-  PatternAdmissionRefusalReasonSchema,
-  PatternAdmissionRefusalSchema,
-  type PatternApplySafety,
-  PatternApplySafetySchema,
-  type PatternBaselineAction,
-  PatternBaselineActionSchema,
-  type PatternCurrentTreeResultClass,
-  PatternCurrentTreeResultClassSchema,
-  type PatternLifecycle,
-  PatternLifecycleSchema,
-  type PatternManifest,
-  PatternManifestSchema,
-  type PatternOwnerTool,
-  PatternOwnerToolSchema,
-  type PatternRecovery,
-  PatternRecoverySchema,
-  type PatternRetirementDecision,
-  type PatternRuleReference,
-  type PatternRuleReferenceInput,
-  PatternRuleReferenceInputSchema,
-  PatternRuleReferenceSchema,
-  type PatternSource,
-  type PatternSourceKind,
-  PatternSourceKindSchema,
-  PatternSourceSchema,
-  type PatternState,
-  PatternStateSchema,
-  type PatternValidationFailureReason,
-  type PatternValidationIssue,
-  type PatternView,
-  PatternViewSchema,
-  type RegisteredPatternManifest,
-  RegisteredPatternManifestSchema,
-  RepoRelativePathSchema,
-} from "./schema.js";
-export {
-  candidateDraftState,
-  candidateUnderReviewState,
-  invalidCandidateState,
-  isAdmittedPatternState,
-  refusedPatternState,
-  retiredPatternState,
-} from "./state.js";
-export {
-  type PatternRulePackReferenceInput,
-  type PatternValidationOptions,
-  type PatternValidationResult,
-  validatePatternManifest,
-} from "./validation.js";
-export {
-  applyAdmission,
-  candidateHandoff,
-  diagnosticAdmission,
-  patternRecovery,
-  patternView,
-} from "./views.js";
