--- tools/habitat-harness/src/service/modules/fix/model/policy/transactions/admission.ts
@@ -1,2 +1,0 @@
-export type { TransactionInputResolution } from "./transaction-input.js";
-export { TransactionInputResolutionSchema } from "./transaction-input.js";
