--- tools/habitat-harness/src/service/modules/fix/model/policy/transactions/index.ts
@@ -1,39 +1,0 @@
-export { runPatternApplyTransaction } from "./apply-transaction.policy.js";
-export { renderPatternApply } from "./render.js";
-export {
-  type DryRunIntent,
-  DryRunIntentSchema,
-  type GritDryRunCommandInput,
-  GritDryRunCommandInputSchema,
-  type LiveWriteIntent,
-  LiveWriteIntentSchema,
-  NonEmptyStringSchema,
-  type PatternApplyDryRunCompletedOutcome,
-  PatternApplyDryRunCompletedOutcomeSchema,
-  type PatternApplyOutcome,
-  PatternApplyOutcomeSchema,
-  type PatternApplyRecord,
-  PatternApplyRecordSchema,
-  type PatternApplyRefusalReason,
-  PatternApplyRefusalReasonSchema,
-  type PatternApplyRefusedOutcome,
-  PatternApplyRefusedOutcomeSchema,
-  type PatternApplyRequest,
-  PatternApplyRequestSchema,
-  parsePatternApplyRecord,
-  parsePatternApplyRequest,
-  parseWorktreeObservation,
-  type RecoveryInstruction,
-  RecoveryInstructionSchema,
-  type ResolvedTransactionInput,
-  ResolvedTransactionInputSchema,
-  resolveTransactionInput,
-  type TransactionInputResolution,
-  TransactionInputResolutionSchema,
-  type TransactionRefusal,
-  TransactionRefusalSchema,
-  UnresolvedTransactionInputResolutionSchema,
-  type WorktreeObservation,
-  WorktreeObservationSchema,
-} from "./schema.js";
-export { observeWorktree } from "./worktree.js";
