+++ tools/habitat-harness/src/service/modules/fix/model/repositories/index.ts
@@ -1,0 +1,1 @@
+export { observeWorktree } from "./worktree.repository.js";
